move thumbnail link generation from SeafConnection to DataManager.
